+++ src/app/loadout-drawer/loadout-utils.ts
@@ -532,6 +532,7 @@ export function getResolutionInfo(
   | {
       hash: number;
       instanced: boolean;
+      bucketHash: number;
     }
   | undefined {
   const hash = oldToNewItems[loadoutItemHash] ?? loadoutItemHash;
@@ -560,6 +561,7 @@ export function getResolutionInfo(
   return {
     hash,
     instanced,
+    bucketHash,
   };
 }
 
@@ -673,6 +675,91 @@ export const isMissingItemsSelector = createSelector(
   (defs, allItems) => (storeId: string, loadout: Loadout) =>
     isMissingItems(defs!, allItems, storeId, loadout)
 );
+
+export const enum FragmentProblem {
+  EmptyFragmentSlots = 1,
+  TooManyFragments,
+}
+
+export const getFragmentProblemsSelector = createSelector(
+  manifestSelector,
+  allItemsSelector,
+  (defs, allItems) => (storeId: string, loadout: Loadout) =>
+    defs?.isDestiny2() ? getFragmentProblems(defs, allItems, storeId, loadout.items) : undefined
+);
+
+function getFragmentProblems(
+  defs: D2ManifestDefinitions,
+  allItems: DimItem[],
+  storeId: string,
+  loadoutItems: LoadoutItem[]
+) {
+  const subclass = getSubclass(defs, allItems, storeId, loadoutItems);
+  if (subclass) {
+    const fragmentCapacity = getLoadoutSubclassFragmentCapacity(defs, subclass);
+    const fragmentSockets = getSocketsByCategoryHashes(
+      subclass.item.sockets,
+      fragmentSocketCategoryHashes
+    );
+    const loadoutFragments = fragmentSockets.filter(
+      (socket) => subclass.loadoutItem.socketOverrides?.[socket.socketIndex]
+    ).length;
+    return fragmentCapacity > loadoutFragments
+      ? FragmentProblem.EmptyFragmentSlots
+      : fragmentCapacity < loadoutFragments
+      ? FragmentProblem.TooManyFragments
+      : undefined;
+  }
+
+  return undefined;
+}
+
+function getSubclass(
+  defs: D2ManifestDefinitions,
+  allItems: DimItem[],
+  storeId: string,
+  loadoutItems: LoadoutItem[]
+): ResolvedLoadoutItem | undefined {
+  for (const loadoutItem of loadoutItems) {
+    const info = getResolutionInfo(defs, loadoutItem.hash);
+    if (info?.bucketHash === BucketHashes.Subclass) {
+      const item = getUninstancedLoadoutItem(allItems, info.hash, storeId);
+      if (item) {
+        return { item, loadoutItem };
+      }
+    }
+  }
+  return undefined;
+}
+
+export function getLoadoutSubclassFragmentCapacity(
+  defs: D2ManifestDefinitions,
+  item: ResolvedLoadoutItem
+): number {
+  // For fragments, we first need to figure out how many sockets we have available.
+  // If the loadout specifies overrides for aspects, we use all override aspects to calculate
+  // fragment capacity, otherwise we look at the item itself because we don't unplug any aspects
+  // if the overrides don't list any.
+  if (item.item.sockets) {
+    const aspectSocketIndices = item.item.sockets.categories.find((c) =>
+      aspectSocketCategoryHashes.includes(c.category.hash)
+    )!.socketIndexes;
+    const aspectDefs =
+      item.loadoutItem.socketOverrides &&
+      _.compact(
+        aspectSocketIndices.map((aspectSocketIndex) => {
+          const aspectHash = item.loadoutItem.socketOverrides![aspectSocketIndex];
+          return aspectHash && defs.InventoryItem.get(aspectHash);
+        })
+      );
+    if (aspectDefs?.length) {
+      return _.sumBy(aspectDefs, (aspectDef) => aspectDef.plug?.energyCapacity?.capacityValue || 0);
+    } else {
+      return getSubclassFragmentCapacity(item.item);
+    }
+  }
+  return 0;
+}
 
 export function isMissingItems(
   defs: D1ManifestDefinitions | D2ManifestDefinitions,